refactor: update routing structure and import paths for authentication and layout components
