+++ app/routes/(public)/license/page.tsx
