+++ app/routes/[handle]/posts/[id]/page.tsx
@@ -8,7 +8,7 @@ import { Button } from '~/components/ui'
 import { dayjs } from '~/libs/dayjs'
 import { getUserPostById } from '~/models/posts'
 import { isAuthenticated } from '~/services/auth'
-import type { Route } from './+types/route'
+import type { Route } from './+types/page'
 
 export const clientLoader = async ({
   request,
