+++ app/routes/auth/google/callback/page.tsx
@@ -1,8 +1,9 @@
 import { href, redirect } from 'react-router'
 import { toast } from '~/components/ui'
 import { authenticateCallback } from '~/services/google-auth'
+import type { Route } from './+types/page'
 
-export const clientLoader = async () => {
+export const clientLoader = async (args: Route.ClientLoaderArgs) => {
   const request = new Request(location.href) // clientLoader の request には hash が含まれないのでここで作る
   const user = await authenticateCallback(request)
   if (user?.handle) {
