+++ app/routes/home/page.tsx
@@ -1,6 +1,6 @@
 import { href, type ClientLoaderFunctionArgs } from 'react-router'
 import { requireUser } from '~/services/auth'
-import IndexPage from './_index'
+import IndexPage from '../page'
 
 export const clientLoader = ({ request }: ClientLoaderFunctionArgs) => {
   const user = requireUser(request, { failureRedirect: href('/') })
