+++ package.json
@@ -40,6 +40,7 @@
     "react-router-dom": "7.2.0",
     "react-twc": "^1.4.2",
     "remix-flat-routes": "^0.8.5",
+    "sonner": "^2.0.0",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.24.2",
@@ -62,6 +63,7 @@
     "remark-frontmatter": "^5.0.0",
     "remark-mdx-frontmatter": "^5.0.0",
     "rollup-plugin-visualizer": "^5.14.0",
+    "rr-next-routes": "^0.0.5",
     "tailwindcss": "^3.4.15",
     "typescript": "^5.7.3",
     "vite": "^6.1.1",
